maintain:sendMail_add_redis_database_make_token_resource
