--- app/Mail/AdminPasswordMail.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace App\Mail;
-
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Mail\Mailable;
-use Illuminate\Queue\SerializesModels;
-
-class AdminPasswordMail extends Mailable
-{
-    use Queueable, SerializesModels;
-
-    public $admin;
-    public $password;
-
-    /**
-     * Create a new message instance.
-     *
-     * @return void
-     */
-    public function __construct($admin, $password)
-    {
-        $this->admin = $admin;
-        $this->password = $password;
-    }
-
-    /**
-     * Build the message.
-     *
-     * @return $this
-     */
-    public function build()
-    {
-        // return $this->view('view.name');
-        return $this->markdown('mail.adminMail', ['name' => $this->admin->name, 'password' => $this->password]);
-    }
-}
