--- app/Policies/PostPolicy.php
@@ -1,94 +1,0 @@
-<?php
-
-namespace App\Policies;
-
-use App\Models\Admin;
-use App\Models\Post;
-use Illuminate\Auth\Access\HandlesAuthorization;
-
-class PostPolicy
-{
-    use HandlesAuthorization;
-
-    /**
-     * Determine whether the user can view any models.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @return mixed
-     */
-    public function viewAny(Admin $admin)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can view the model.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @param  \App\Models\Post  $post
-     * @return mixed
-     */
-    public function view(Admin $admin, Post $post)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can create models.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @return mixed
-     */
-    public function create(Admin $admin)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can update the model.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @param  \App\Models\Post  $post
-     * @return mixed
-     */
-    public function update(Admin $admin, Post $post)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can delete the model.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @param  \App\Models\Post  $post
-     * @return mixed
-     */
-    public function delete(Admin $admin, Post $post)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can restore the model.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @param  \App\Models\Post  $post
-     * @return mixed
-     */
-    public function restore(Admin $admin, Post $post)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can permanently delete the model.
-     *
-     * @param  \App\Models\Admin  $admin
-     * @param  \App\Models\Post  $post
-     * @return mixed
-     */
-    public function forceDelete(Admin $admin, Post $post)
-    {
-        //
-    }
-}
