+++ resources/views/mail/adminMail.blade.php
@@ -1,5 +1,5 @@
 @component('mail::message')
-    hello {{$name}}
+    hello {{$email}}
     'you are successfully registered as an admin
     you must use the following password to login
     password : {{$password}}
