+++ routes/api/admin.php
@@ -19,5 +19,5 @@ use Illuminate\Support\Facades\Route;
 Route::post('admin/login',[LoginController::class, 'adminLogin'])->name('adminLogin');
 Route::group(['middleware' => ['auth:admin']], function () {
     Route::apiResource('admins', AdminController::class);
-    
+    Route::get('admin/permissions', [AdminController::class,'getPermissions']);    
 });
